+++ plotoptix/_load_lib.py
@@ -289,6 +289,17 @@ def load_optix():
     optix.set_max_accumulation_frames.argtypes = [c_int]
     optix.set_max_accumulation_frames.restype = c_bool
 
+    optix.encoder_create.argtypes = [c_int, c_int, c_int]
+    optix.encoder_create.restype = c_bool
+
+    optix.encoder_start.argtypes = [c_wchar_p, c_uint]
+    optix.encoder_start.restype = c_bool
+
+    optix.encoder_stop.restype = c_bool
+
+    optix.encoded_frames.restype = c_int
+    optix.encoding_frames.restype = c_int
+
     optix.set_gpu_architecture.argtypes = [c_int]
 
     optix.set_library_dir.argtypes = [c_wchar_p]